Move private api to it own file, made vpce config optional & removed read artifact
