+++ deployment/aws/cdk/config.py
@@ -86,6 +86,6 @@ class StackSettings(BaseSettings):
     max_concurrent: Optional[int] = None
 
     # The VPC Endpoint ID
-    vpc_endpoint_id: str = "{{VPCE_ID}}"
+    vpc_endpoint_id: None | str = None
 
     model_config = SettingsConfigDict(env_prefix="TITILER_STACK_", env_file=".env")
