--- deployment/aws/cdk/read.14868.1.lock
@@ -1,1 +1,0 @@
-14868
